[test] Shared assertions for pattern form
